src/promo-agent/requirements.txt
Flask==2.2.2
gunicorn==20.1.0
requests==2.28.1
schedule==1.1.0
python-dotenv==0.21.0
